+++ pkg/commands/typed_command.go
@@ -2,7 +2,7 @@ package commands
 
 import (
 	"context"
-	"github.com/matzefriedrich/cobra-extensions/pkg/reflection"
+	"github.com/matzefriedrich/cobra-extensions/internal/reflection"
 	"github.com/matzefriedrich/cobra-extensions/pkg/types"
 	"github.com/spf13/cobra"
 )